make contentversion a float to support future updates (e.g. CK version 5.1)

+++ src/sdk/PnP.Core/Model/SharePoint/Pages/Internal/RtePageSettings.cs
@@ -5,6 +5,6 @@ namespace PnP.Core.Model.SharePoint
     internal sealed class RtePageSettings
     {
         [JsonPropertyName("contentVersion")]
-        public int? contentVersion { get; set; }
+        public float? contentVersion { get; set; }
     }
 }
